refactor(framework): Isolate warm AgentApp exec handoff
